pyproject.toml
[tool.poetry]
name = "ED Control Point"
version = "0.1.0"
description = ""
authors = ["Dipping Dino <diction.looms-6z@icloud.com>"]
readme = "README.md"

[[tool.poetry.packages]]
from = "src"
include = "*"

[tool.poetry.dependencies]
python = "^3.10"
rangehttpserver = "^1.4.0"
requests = "^2.32.3"
duckdb = "^1.3.0"
boto3 = "^1.38.46"
python-dotenv = "^1.1.1"


[tool.poetry.group.dev.dependencies]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
